Allow createType to be used for composite types (#74)
If the `options` passed to `createType` isn't an array treat it as a set
of attributes for a composite type. This is the only form of type
creation that conforms to the SQL standard.

This does leave creation of range and base types unsupported, and may
make it more difficult to add support for those in the future. But,
those types would generally also require defining functions which is
also not supported by a helper; so anyone wanting to create those types
is likely to need to use raw SQL anyway.