Add --single-transaction option
This adds a new boolean option, `--single-transaction`. When true, all pending migrations are combined into a single database transaction. This is useful if you want to make sure your database will get either _all_ of the pending migrations or _none_ of them, without ending up in an indeterminate state in between.