Implement failsafe locking (#239)
This switches to PostgreSQL's advisory locking. Unlike the current strategy of using a table comment, it cannot leave a dead lock lying around, even if node-pg-migrate crashes. This is an implementation of what I suggested in #232.

I removed the manual `unlock` command because there's nothing for it to do. The only way to get stuck is if you have an actual node-pg-migrate process that is connected to the database and hung, in which case you should just kill it and that will release the lock.